+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add `unregister_device` to unregister a device, allowing to register it again. 
+
 ## [1.0.0-rc.0] - 2021-05-05
 ### Changed
 - Standardize subscriptions (see